+++ tests/CrawlerProcess/caching_hostname_resolver_ipv6.py
@@ -15,7 +15,7 @@ if __name__ == "__main__":
     process = CrawlerProcess(
         settings={
             "RETRY_ENABLED": False,
-            "DNS_RESOLVER": "scrapy.resolver.CachingHostnameResolver",
+            "TWISTED_DNS_RESOLVER": "scrapy.resolver.CachingHostnameResolver",
         }
     )
     process.crawl(CachingHostnameResolverSpider)